# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vocab System API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /api/learning-records:
    get:
      parameters:
        - in: query
          name: studentId
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200': { description: OK }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recordType, payload]
              properties:
                studentId: { type: string }
                vocabularyId: { type: string }
                recordType: { type: string }
                payload: { type: object }
                score: { type: integer }
      responses:
        '200': { description: Created }
...